move price_info to its own crate
