+++ ethcore/Cargo.toml
@@ -45,6 +45,7 @@ lru-cache = "0.1.0"
 native-contracts = { path = "native_contracts" }
 num = "0.1"
 num_cpus = "1.2"
+price-info = { path = "../price-info" }
 rand = "0.3"
 rlp = { path = "../util/rlp" }
 rust-crypto = "0.2.34"
