+++ ethcore/src/lib.rs
@@ -99,6 +99,7 @@ extern crate lru_cache;
 extern crate native_contracts;
 extern crate num_cpus;
 extern crate num;
+extern crate price_info;
 extern crate rand;
 extern crate rlp;
 extern crate rustc_hex;
